requirements.txt
streamlit
langchain==0.3.0
pandas==2.2.2
numpy==1.26.4
docker
python-dotenv
json-repair
nltk
transformers==4.44.2
streamlit-agraph
neo4j
torch==2.2.2
faiss-cpu
tiktoken
langchain-community
sentence-transformers
requests
PyPDF2
spacy==3.7.2
scispacy==0.5.4
openai>=1.40.0
tqdm